Created the minfx.grid.grid_split_array() function.
This is similar to the minfx.grid.grid_split() but it instead works with the output of the
minfx.grid.grid_point_array() function, i.e. it accepts a list of grid points instead of generating
the points itself.  The function is used to split up a grid search into equal parts for use in
parallelisation.